requirements.txt
streamlit==1.20.0
pymongo==4.2.0
streamlit-option-menu==0.3.1
pandas==1.5.3
boto3==1.24.6
opencv-python==4.7.0.72
Pillow==9.3.0
python-dotenv==1.0.0
twilio==7.14.0
plotly==5.10.0
Ultralytics
Opencv-python
Imagelib